Sub-directory Log Files
This example is a log file that will keep track of one years worth of logging and be found in the folder 2020. In this case, I cite an example of one incident that takes place on January first 2020, with an image file in folder 2020/01/01/incident-one.jpg, but to keep the links functional the root domain needs to be used to identify the page and the image correctly, so it is written with the domain name: example.com. Simply replace example.com with your own domain, and when you call the link (<a href="http://example.com;/datetime/2020/01/01/incident-one.jpg">, it will display the image that you've place there. Relative tags won't work in deep directories. We have to use an absolute location to make the pages functional. This also applies to the css for the page, which in this case, comes from the W3 Schools.